Move OTel wrappers to the modules defining corresponding interfaces
Expose ObservingPersist and ObservingVersionStore in the modules that
define their respective interfaces.

Keep Quarkus CDI beans in the quarkus-common module (extending the moved
classes).
