+++ servers/quarkus-common/src/main/java/org/projectnessie/quarkus/providers/versionstore/QuarkusObservingVersionStore.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) 2023 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.quarkus.providers.versionstore;
+
+import jakarta.inject.Singleton;
+import org.projectnessie.quarkus.providers.NotObserved;
+import org.projectnessie.versioned.ObservingVersionStore;
+import org.projectnessie.versioned.VersionStore;
+
+/** CDI bean for {@link ObservingVersionStore}. */
+@Singleton
+public class QuarkusObservingVersionStore extends ObservingVersionStore {
+  public QuarkusObservingVersionStore(@NotObserved VersionStore delegate) {
+    super(delegate);
+  }
+}
